+++ src/shared/enums/component-id.enum.ts
@@ -4,8 +4,10 @@ export enum ComponentId {
     SEEDED_LIST_INPUT = 'seeded-list-input',
     SEED_MODAL = 'seed-modal',
     WEEKLY_DAY_SELECTOR = 'weekly-day-selector',
+    WEEKLY_PARTICIPANTS_BULK_CONFIRM_FREESTYLE_BUTTON = 'weekly-participants-bulk-confirm-freestyle-button',
     WEEKLY_PARTICIPANTS_BULK_CONFIRM_INPUT = 'weekly-participants-bulk-confirm-input',
     WEEKLY_PARTICIPANTS_CONFIRM_MODAL = 'weekly-participants-confirm-modal',
+    WEEKLY_PARTICIPANTS_CONFIRM_SELECTOR = 'weekly-participants-confirm-selector',
     WEEKLY_PARTICIPANTS_DECONFIRM_SELECTOR = 'weekly-participants-deconfirm-selector',
     WEEKLY_PARTICIPANTS_HISTORY_DATES_SELECTOR = 'weekly-participants-history-dates-selector',
 }