+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 2024-01-18 - 0.2.0
 
+-   Update bulk-confirm command to support discord user select
 -   Fix seed command
 
 ## 2024-01-11 - 0.1.2